requirements.txt
google-auth==1.24.0
google-cloud-bigquery==2.6.1
requests==2.25.1
